COPYING: new file
Open-Sankoré is released under the terms of the GNU Library Genereal
Public License, version 2, with a specific linking exception for the
OpenSSL project's "OpenSSL" library, so state that as in a general GNU
project and remove everything which is useless (and also misleading).

The LGPL-2 version is taken from the Debian base-files_6.9 package,
file /usr/share/common-licenses/LGPL-2.  The OpenSSL exception comes
from the GNU Wget license as in the Debian wget_1.13.4-3 package, file
/usr/share/doc/wget/copyright.