Removed export of polygons outside of groups
The lines removed in this commit led to (presumably) unintended
behavior: when saving a document, pen strokes were saved not as a group,
but just as their constituent polygons.
This meant that the following block of code, that handles saving pen
strokes correctly, was never executed.

This doesn't fix the z-Value issue (#12), but it seems to be a step in the right
direction.