Check settings at launch to avoid crashing
Some settings were changed between v1.02 and 1.10 (current), and some
of these changes cause OpenBoard to crash at launch. This commit adds
a function to check for these specific new settings, and wipe the old
values if they are found.
This avoids problems when the user upgrades from 1.02 without deleting
their configuration file.

(This is an alternative to having a post-install script, which would be
ineffective in a multi-user configuration)