Correct screen indexing in some cases
Several issues remain with multi-screen mode on Linux. The behavior is
inconsistent from one desktop evironment to the next, making it hard to
work around these problems. Among the known issues at this stage:

On Ubuntu 14.04, a call to QWidget::setGeometry requires the widget to
be hidden on KDE, but visible on MATE, for the geometry changes to take
effect.
Despite the widget's geometry being updated by this call, the windows
aren't necessarily moved. Meaning that the control and display widgets
will tend to be displayed on the same monitor, even though their
positions are correctly set to different areas on the extended screen.

In the current state, this behavior is observed on MATE. Unity works
fine and KDE only has transient positioning issues (for example,
swapping control and display windows in multi-screen mode leads to both
windows being placed on the same monitor, until multi-screen mode is
turned off then on again).

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch dev
# Your branch is ahead of 'origin/dev' by 29 commits.
#   (use "git push" to publish your local commits)
#
# Changes to be committed:
#	modified:   src/core/UBApplicationController.cpp
#	modified:   src/core/UBDisplayManager.cpp
#	modified:   src/core/UBDisplayManager.h
#
# Changes not staged for commit:
#	modified:   release_scripts/linux/build.sh
#	modified:   release_scripts/linux/package.sh
#
# Untracked files:
#	release_scripts/linux/generateDependencies.sh
#