Removed blending of adjacent polygons in marker strokes
Due to antialiasing, adjacent polygons are separated by a very fine
space. The previous solution attempted to hide this by adding a border
to the polygons. The border of adjacent polygons would overlap, which
was visible (despite the attempted color correction) and, more
importantly, caused massive lags especially on Linux.

Therefore it has been removed but feel free to revert this commit some
day and try to fix this more cleanly.