Added stroke simplification algorithm
If enabled in the preferences menu, pen and marker strokes will be
replaced by a simplified stroke after they are drawn.

The algorithm is very basic (for now): if three points are almost lined
up (the threshold angle can be specified in the config file), then the
middle one is removed. This is repeated over the whole stroke; new
polygons are then generated based on the simplified stroke points.

This typically cuts down on number of points and polygons by a factor of
about 10, while having minimal visual impact.