Fixed PDF export page size
In some cases, export of a document containing a PDF background to PDF
caused the contents to be truncated.

The "simple" PDF exporter will now set the output page size to be equal
either to the document nominal size or, if the document has a background
PDF item, to the size of this item.