Fix for video-related crashing on Windows
Issue observed was that OpenBoard would crash on some Windows systems
when a video was on the page and that page was saved (due to switching
to document mode, auto saving, or duplicating the page), or when cutting
the video with Ctrl-X.

This was due to QTBUG-32522, where setting the video output for a
QMediaPlayer that is hidden results in a crash.

This commit is a work-around for this Qt issue, and so should be reverted
if and when the upstream issue is fixed.