Allow cut-and-pasting items between pages
Previously, items such as audio, video, and widget items could not be
cut and pasted because when the page was saved, any item that had been
deleted was permanently removed (including its source file, in the case
of these mulitmedia items).

This commit prevents the deletion any item in the undoStack
(i.e, items that have been deleted) when the document is persisted,
meaning the source file is still available when pasting the item on
another page (or document).

Note that this can lead to cases where the source file is kept in the
document even when no item is present. This should be a relatively rare
occurence, however.