Update podcast.pri
When trying to compile on Ubuntu 20.04, I found the following bug and suggestion for improvement.
(bug) The "&" needs to be removed, as the output of the "ffmpeg --version" command is piped to the "grep" command.
(improv) Running "qmake" with With option "-Wall" also shows that using an un-escaped "\" is deprecated. It needs to be escaped with a "\".