feat: Use pkg-config for poppler and quazip on Linux
In particular the quazip library, tied in with QT as it is, can be found
in several different locations depending on the QT version, how the
distro packages QT, etc. Use the PKGCONFIG variable for adding libraries
and includes, as pkg-config automatically deals with different
locations.

Additionally, the quazip library has several different names, which
means that directly including the standard name doesn't always work.
Instead use pkg-config to check under which name it is available on the
system.

In case that quazip isn't registered within pkg-config, also check
whether the old paths work and add them manually. This is to ensure
complete backwards-compatibility.

Finally, if everything fails to find quazip, the build is aborted with
"quazip not found".

Taken with small changes (mostly QT6 support) from
https://github.com/OpenBoard-org/OpenBoard/pull/651/files#diff-22f2374aefacf9d7887f143c05ae203e30e2b2f0659368b67d5472c6aaa55731R473-R498