feat: Clean up linked libraries for podcast on Linux
podcast on Linux specifies a lot of unnecessary packages which are
either not needed anymore or transitive dependencies of other libraries.
The remaining libraries are now linked using PKGCONFIG, which is cleaner
and allows for flexibility when the libraries are packaged differently.

Removed aswell is the check for FFMPEG 2.8. This particular condition
replaced swresample (of ffmpeg) with avresample (of libav). The history
of this issue is quite old and basically boils down to Ubuntu 14.04
shipping with libav which, at the time, also provided an ffmpeg command.
This libav-ffmpeg used the 'av'-version of swresample.
By Ubuntu 15.04, libav didn't provide ffmpeg and ffmpeg was the actual
command everyone expects. Every release since then is unaffected.

On a side note, 14.04 is an LTS that's technically still supported by
Canonical until April 2024. In conclusion, to be affected on a
supported configuration, one must
1. use Ubuntu 14.04 (less than 5 months till EOL),
2. use a very old libav which provides ffmpeg (now defunct),
3. update to OpenBoard 1.7 (the newest version).

I think this configuration is very unlikely to happen and it's better
to clean this up all at once.

Inspired by
https://github.com/OpenBoard-org/OpenBoard/pull/651/files#diff-536f069bceda9079731708179068a3a26ad5306893ef8744f61e296cce839fe2R60