fix: Make ruler background rounded
The ruler widget has several gradients as a background. At the top and bottom,
the background has an edge color, while in the center, it has a different color.
One gradient linearly fills from the top to a quarter, a second from the bottom
to three quarters, while the two quarters in the middle are the center color without
gradient.

The previous approach constructed 3 rectangles, one for each gradient change,
to build the background. This prevented the use of rounded corners as the borders in
the middle are straight, not rounded. The disadvantage is that the square corners of
the background overlap the rounded corners and "overflow" the ruler.

Instead, the new approach creates a single gradient and a path, and tells Qt to fill
the path with the gradient. The gradient has one stop at the top, one at a quarter, one at
three quarters and one at the bottom. The two stops in the middle have the same color, so
effectively no gradient.
The path can also be reused to draw the border. This ensures that the border and the
background always have the same outline.