refactor: Remove unused function call
Git history doesn't go far enough to determine why this call exists;
it's at least 13 years old. This section of code hasn't ever changed.
It's reasonable to assume that it's a typo (maybe first implementation
attempt) and the compiler just didn't warn about it, so it was
overlooked.
Now the compiler warns about it.