refactor (UBSettings): Make proxy credentials nonpersistent
The proxy credentials are currently stored in an insecure way, and
should not be stored persistently if it can be avoided.

For architectural reasons, the credentials have to be saved for the
runtime of the program to be accessed by the network manager.

Clear the credentials at the end of the program.