refactor(UBPersistenceManager): Make createDocumentProxyStructure static
The method `createDocumentProxyStructure` doesn't depend on the object,
so can be made static. QtConcurrent::mapped requires the argument of
the method to be const, since it creates a new list for the mapping.

The lambda function marks its arguments as const by default. Since
the method is now static, the lambda can be replaced with a direct
reference.