refactor: Remove usage of QT_IMPLICIT_QCHAR_CONSTRUCTION on Windows
The macro `QT_IMPLICIT_QCHAR_CONSTRUCTION` allows converting an
integer into a QChar.
Starting with 6.0, the default behaviour of Qt is to only allow
constructing a QChar from an integer explicitly, but until 6.8,
above macro allowed the implicit conversion. Qt 6.8 removes the
macro, requiring explicitly calling a QChar constructor when
creating a QChar from an integer.