refactor: use different setContent overload
The old version of setContent returns bool and takes in pointers
to allow access to the error information. This was replaced in Qt 6.5
with a ParseResult struct which allows boolean checking and direct
access to the information without additional pointers.

The old version had separate overloads to specify whether to use
namespace processing or not. This, among other options, can now be
specified via ParseOptions which were introduced in Qt 6.5 too.