feat: User-defined podcast fps
Currently, the podcast recording is set to a fixed 10 fps. While
there exists a setting to allow for customisation, it remains
unused.

This change takes the user setting into account. The actual fps
are determined by the activity on the board, so it represents
a maximum fps. This value is further refined depending on the
selected podcast preset:
- Full:   Use the value directly
- Medium: The max. fps are halved
- Small:  The max. fps are divided by 4

The default has been bumped up to 100 fps. This value is never
actually reached - the resulting fps hover mostly around 40-70.
During testing on my rather powerful machine (with values exceeding
250 fps), the observed frame rate never got appreciably higher than
100 fps (maximum was 127 fps for short periods at 1000).
Thus, the Full preset in its default configuration captures all
activity.
Choosing a lower preset decreases the fps accordingly, with 25 fps
at the lower end of what is perceived as fluid motion, though actual
framerate can be around 12-18. Further tuning is possible via the
"Podcast->FramesPerSecond" setting.